[#5157] Return dataObjInfo_t in rsPhyPathReg
Instead of using two separate functions for registering intermediate
replicas, we should go back to using rsPhyPathReg and use a different
mechanism for obtaining the resulting data object info from a
registration.

Returns to using rsPhyPathReg for registration of intermediate replicas
in rsDataObjOpen. The dataObjInfo_t structures populated for registering
a data object or a replica now replace the dataObjInfo held by the open
L1 descriptor.

Adds a new function called irods::find_l1desc which takes a logical path
and a resource hierarchy and tries to match the data object information
of an opened L1 descriptor. If found, a pointer to the L1 descriptor is
returned.

Also adds the ability to release a pointer held by a lifetime_manager.
This causes the lifetime_manager to surrender control of the managed
object to the caller, who is then made responsible for freeing the
memory therein.

Refactored some pieces of rsPhyBundleColl and rsChkObjPermAndStat to
replicate to cache. Removed replDataObjForBunOpr because its existence
added very little in the way of code readability/reusability.