[#3049] Server stores rule execution info in catalog as JSON.
- Added new database column (EXE_CONTEXT) which is responsible for
  holding the REI context information.
- The delay server now migrates REI files into the catalog following
  rule execution.
- Fixed comment in replica_access_table implementation.
- Fixed delay server zombification issue by moving the retrieval of a
  rule's catalog information to the thread pool task.

Session variables are not allowed in delay rules at all (this includes
comments). Attempting to schedule a delay rule with session variables
will result in an error. Session variables can be used indirectly by
capturing their value and using the new name for the value.

Existing delay rules that use session variables will not be migrated
into the catalog. Instead, they will continue to depend on the REI file
until the administrator removes the dependency on the session variables.

For more details, see the following RFC:

  https://github.com/irods/irods_rfcs/blob/master/0006_static_peps_and_delay_rules.md