[#3848] Add replica_state_table Singleton
Adds a global Singleton a la replica_access_table, but is specific to
individual agents. It is like an L1 descriptor which keeps track of all
information for each replica of a given data object. It consists of a
map which keys on logical paths and has a value consisting of the state
of the data object in the catalog at time of open (known as the "before"
state) and an updating state which will serve as the input to a call to
data_object_finalize (known as the "after" state). The table takes on
the following form:

{
    "/full/logical/path/a": {
        "data_id": <int>,
        "replicas": [
            {
                "before": {
                    <r_data_main_column>: <string>,
                    ...
                },
                "after": {
                    <r_data_main_column>: <string>,
                    ...
                }
            },
            ...
        ]
    },
    ...
}

The "before" state exists in order to keep track of the replica status,
size, checksum, size, and other information before any changes were made
due to data movement. The "after" state exists to serve as input to a
catalog update on the data object once the operation has completed.

Currently only supports one open per agent. If an operation needs
additional replicas to be opened at the same time (e.g. replication and
phymv), the operation should keep track of its own information as
needed.