[#5179] Fix Test_Resource_Replication_With_Retry
This change properly handles errors in rsDataObjRepl which come from
rsDataObjClose. The error codes are now returned which allows the retry
mechanism of the replication resource to do its job.

Replica access table restores the entry for an opened replica on failure
in order to maintain consistent client/server state. However, the
replica token is lost inside of rsDataObjRepl and a retry will fail due
to a missing replica token keyword.

There was also an issue where rsDataObjClose was not marking replicas as
stale on failure. This has been amended in order to allow for retries on
failed replications.