[#5149] New replica and data_object_proxy features
Adds a free function the replica library which retrieves a replica's
status from the catalog.

Adds a free function to the irods::experimental::data_object namespace
which duplicates a data_object_proxy (including its underlying
DataObjInfo struct/linked list). Returns a data_object_proxy and a
lifetime_manager for the underlying struct.

Adds a free function to the irods::experimental::data_object namespace
which finds a replica in a data_object_proxy::replica_list_ by searching
for a replica with a matching resource hierarchy.

Adds a member function to the data_object_proxy which allows for adding
a new replica to the linked list. The interface for the underlying
std::vector of replica_proxy objects will be changed in the future for a
more consistent experience, but this is needed in order to add new
replicas for the purposes of finalizing in logical locking.