[#5184] New check_type EMPTY_STDOUT in command.check_command_output
My test for #5184 needs to be able to check that stdout is empty while
ignoring anything that might be in stderr. At present, the only check_type
that checks for an empty stdout is EMPTY, but check_command_output will
return false if anything at all is in stderr and EMTPY (or any of the
non-STDERR values) is passed in for check_type.

This commit introduces the EMPTY_STDOUT check_type value, which checks
for an empty stdout and ignores stderr completely.