[#3848] Refactor rsDataObjClose with proxies
Logical locking will change the way rsDataObjClose addresses replicas in
the catalog by updating the in-memory object model and replica state
table rather than directly manipulating the catalog. This change
refactors much of rsDataObjClose to make the boundaries clearer so that
changes can be made more cleanly for logical locking.