[#5149] New replica lib functions
Adds a duplicate_replica method to the replica namespace in
replica_proxy. Similar to duplicate_data_object.

Also fixes duplicate_data_object. DataObjInfo has a few sub-structs
which need to be allocated and copied for each duplicated node in the
linked list. Some items are not in use and are not touched by the proxy
object.

Adds get_replica_size_from_storage which allows specifying a logical
path and a replica number or leaf resource name to find the size of a
physical replica representation in the storage medium.