[#5213] unit_test test_path: compare strings of incompatible path types
clang 10 throws an error because it can't figure out how to compare
an std::filesystem::path and an irods::experimental::filesystem:path

This error doesn't come up with clang 6, so either the
__cpp_lib_filesystem ifdef isn't triggered, or there's some implicit
type coercion happening. If it's the latter, I haven't the faintest
idea what types these objects are being coerced to.