[#5263] Fix size/checksum mismatch handling
Size and checksum mismatch handling as it relates to replication was not
behaving exactly as it should. The dataSize/bytesWritten are used by
rsDataObjClose to determine if a replica has the correct size when
validating the replica on close.

Sets openFlags differently in copy and repl so that they use:
O_CREAT | O_WRONLY | O_TRUNC.