[#5251][#5252] Fixed the checksum API.
The checksum API now has two modes: verification and lookup/update.

Verification mode performs the following operations:
1. Reports replicas with mismatched size information (physical vs catalog).
2. Reports replicas that are missing checksums to the client.
3. Reports replicas with mismatched checksums (computed vs catalog).
4. Reports if the replicas do not share the same checksum.

Step 3 can be time consuming depending on the size of the replica. The
server can be instructed to skip detection of mismatched checksums by
passing the NO_COMPUTE_KW keyword.

The verification operations work for one or more replicas. However, when
a specific replica is targeted, step 4 is not performed.

Lookup/Update mode performs the following operations:
1. Returns the existing checksum or computes and updates checksums.
2. Reports if the replicas do not share the same checksum assuming no
   errors were encountered during the previous step.

The following keyword(s) are now NOPs (in regards to rxDataObjChksum):
- VERIFY_VAULT_SIZE_EQUALS_DATABASE_SIZE_KW

Operations that target multiple replicas will only affect replicas that
are marked good. This means intermediate, locked, and stale replicas
will be ignored.

Operations that target a specific replica are allowed to operate on
stale replicas.