[#5018] Stale un-closed replicas on agent teardown
When a replica has not been closed, agent teardown logic will come along
and close all of the open L1 descriptors. In addition to this, the agent
teardown process should attempt to finalize these replicas in the
catalog and mark the replicas as stale. The following operations are
performed:

  - check size in the vault and update catalog
  - perform checksum on replica as determined by existing catalog
information and L1 descriptor keywords
  - update modify timestamp for open-for-write cases
  - set replica status for stale

In the open-for-read case, the replica status is not touched as nothing
moved.

This commit also correctly sets the status for replicas opened for write
with no bytes written by restoring the replica status recorded at the
time of open before it was moved to the intermediate status.