[#5314] Update replica status before opening
When opening a replica for write, the replica status changes to
intermediate to indicate to other agents that the replica's data is not
at rest and should not be opened for read or write. The replica status
should be set to intermediate before the actual open is invoked to
ensure that no other actor can slip in between the open and the update
to the catalog and open the file.

This change also fixes an issue with rsDataObjPhymv. The source replica
was being opened with O_RDWR when in reality it should be opened with
O_RDONLY. However, the use of O_RDWR was implicitly enforcing a
constraint on phymv which disallows the operation if the source replica
is found to be read-only. This would mean that the replica cannot be
unlinked at the end of the operation and the catalog would lose track of
the data (potentially leaking data). Before attempting the phymv
operation, the API now attempts to open the source replica for write
using rsFileOpen/rsFileClose. If successful, the operation proceeds. If
not, the operation aborts with the Permission Denied error returned by
the OS.