[#5149] Improvements for proxy classes
Adds helper types for the replica_proxy and data_object_proxy classes so
that they are easier to use. Up to this point, DataObjInfo was still
easier to use in an API declaration due to the simplicity of the type
name. This allows use of proxies directly in various interfaces.

Creates a compilation unit for key_value_proxy to reduce the amount of
code in the header file. Only contains one non-template function at this
point in time.