[#3848] Add RST interfaces/documentation/tests
As part of the effort to implement logical locking, we need the ability
to easily and atomically update the status of all replicas for a given
data object. This change adds several new interfaces including a
"publish" function in order to stamp out a replica state table entry to
the catalog as part of the finalize workflow.

This change also allows multiple references to the same replica state
table entry for things like replication. An internal reference count is
maintained by the replica state table so that multiple opens can track
replica states at the same time.