[#4331] Add some utilities in support of finalize
Adds a convenience function to call replica_close without touching the
catalog which is useful for internal APIs. These will use finalize and
update the catalog in a way that makes sense for the operations they are
supposed to perform.

Adds some new interfaces to irods::file_object and irods::data_object
for use in the voting process.