[#4331] Add file_modified for data_object_finalize
Adds support for a file_modified entry which is attached to the JSON
representation of a particular replica. At this time, the first replica
to appear with a file_modified key will be selected for the fileModified
plugin operation. The file_modified key in the overall structured input
to data_object_finalize signals to the API whether fileModified should
be invoked, regardless of the presence of file_modified in the replica
entries.

Before triggering fileModified, any replica state table entries for the
target data object are removed once the API has published the input to
the catalog. This holds a new requirement that replicas should be closed
before calling data_object_finalize if fileModified is to be invoked.
This would affect things like replication and physical moves of replicas
which open multiple replicas of the same data object. The source replica
should be closed before closing the destination replica so that
fileModified can be invoked safely.

Adjusts replica_close to only erase replica state table entries when
finalizing as data_object_finalize will be used more internally for
finalizing data objects.