[#3490][#5312] Add new rules for repl and phymv
rsDataObjRepl and rsDataObjPhymv now have a new set of established rules
defining when replication is and is not allowed and what happens as a
result of attempting replication in those scenarios. This change adds a
large set of tests for each possible scenario given leaf resources and
different modes of operation for each API (as reflected and executed in
their respective icommands, irepl and iphymv).

rsDataObjPhymv was refactored a bit to resemble rsDataObjRepl. Much of
the functionality of one can be found in the other and may be moved into
a library at a later time.