[#5195] Adjust open/close for new RST interfaces
Adjustments made to rsDataObjOpen and rsDataObjClose due to changes in
how the replica state table works. Allows for multiple replicas to be
open on the same data object for things like replication. Fencing in
special collections and bundle resource codepaths so that they do not
use intermediate replica states.

Change source/destination close order in copy so that it can use
finalize in the future.

Adjusts rsDataObjGet to work with some of the changes in rsDataObjOpen.

Also populates expiry_ts in database plugin at registration so that the
information is in an expected state.

Performs the following operations on agent teardown for all open L1
descriptors:

 - Size calculation
 - Checksum calculation
 - Stale replica

...and updates the catalog via rs_data_object_finalize.