[#4010] Favor stale replica in write vote
This allows replication from a leaf resource found in a replicating
hierarchy to its own parent to update stale replicas found in the
hierarchy. This does not allow creating missing replicas.

This change also returns all replicas in the result set for
getDataObjInfo even when the REPL_NUM_KW is used.

This change removes the filter on REPL_NUM_KW in getDataObjInfo so that
all replicas are returned in the result set even if a specific replica
is being targeted. getDataObjInfo should not limit itself to the replica
number - the caller should be responsible for finding the needed replica
number, or an equivalent General Query should be used.

A number of APIs and some pieces of the voting mechanism are leaning on
this behavior, so they were adjusted to work with this change. The
compound resource in particular needed to rework its voting in the
open_with_prefer_archive behavior.