[#2759] Delay server now schedules rules for execution based on their priority.
Valid priority levels are defined by the following range: 1 - 9

The delay server gathers and schedules rules in descending order of
priority level.

<PRI> tag is no longer supported by delay(). Use <PRIORITY> instead.
Attempting to use <PRI> will result in an error.

Reoccurring rules that have an empty priority level will be updated to
have a priority level of 5 following successful execution. This also
applies to rules that have an invalid priority level.