[#3848] Add logical_locking library
This is just the library for changing replica statuses via the
replica_state_table. Adds the ability to lock/unlock a data object,
specify the desired status of the target replica as well as its
siblings, and uses the data_status column to store previous replica
status information. This will be needed in the future for read locks as
multiple readers will lose the information if it's not stored in the
catalog.