[#5542] Lock objects before creating new replicas
When a new replica is being created on an existing data object, the data
object should be locked before attempting to create the new replica.
This prevents a small but significant hole in the ILL protection which
would allow two agents to create new replicas on the same data object at
the same time.

To achieve this, two new functions were added to the ILL interface which
do not require a replica number or replica status for locking/unlocking.
The statuses for all of the existing replicas are locked or unlocked
appropriately with any mention of a target replica being ignored in the
implementation as there is no target replica for this use case.

This change also fixes the lock-checking logic in rsDataObjOpen. First,
the replica targeted for open is checked to see if it is in the
intermediate state and, if so, whether the agent attempting the open is
authorized to open it (i.e. replica token). If the replica is at rest,
the other replicas for the data object are checked for a lock status. If
any replica is locked, the data object is considered locked and the open
is not allowed to continue.