[#5187][#5422] Stale sibling replicas on write
When replica_close is finalizing a newly created replica or a replica
which has been overwritten, sibling replicas should be marked as stale
because the newest replica reflects a potentially different truth from
that of the existing replicas. As part of this effort, replica_open has
been modified to handle failures from rs_get_file_descriptor_info a bit
more gracefully by unlocking the replica which was opened and unlinking
in the event that the open was meant for a create.

Add istream test for sibling replica marked stale on overwrite and
create.

Also makes sure that all exceptions in rs_get_file_descriptor_info are
caught.