[irods/irods#5360] Return more detailed codes in iphymv
This change introduces more exit codes to iphymv based on a family of
error codes returned from the server in rsDataObjPhymv. Doing so allows
system implementers to take action based on a number of common use cases
such as retrying in the face of connection/network errors versus the
target replica being written to a node with a full disk.