[#5360] Add data transfer failure tests
This change uses mungefs to create a series of tests which demonstrate
failures in data transfer and how various API endpoints handle these
situations. In particular, iphymv has been changed to emit different
error codes for different types of errors, so these tests help ensure
that the error codes are propagating correctly. The other API endpoints
are also now being tested to ensure that the replicas end up in the
correct locations and with the correct replica statuses.

Also added some sleeps to the test_rc_data_obj C API test which asserts
some things which only occur after agent teardown which may not be
reflected in the catalog immediately after disconnecting.