[#5554] Add test for reg_data_obj dynamic post-PEP
The dataObjInfo struct was not being fully populated when registering a
data object in the database. In particular, the user name, zone name,
and create time were not populated so PEPs which look inside these
structs were not able to extract the appropriate information. This has
been fixed but needed a test which this change provides.