[#4236][#5421] Add tests for ILL in various APIs
Adds tests for enforcement of ILL in various APIs including
rcDataObjUnlink, rcDataObjTrim, and rcDataObjRename, none of which
actually open replicas. Also rearranged some of the tests for reasons of
efficiency and clarity.

Adds an unlocking mechanism via rcModDataObjMeta in the
test_replica_open_and_close tests which ensure that cleanup can occur on
replicas which did not have their statuses updated on purpose as part of
the test.