[#5421] Add ill::try_lock
This function was added to ILL as a general way of checking for locked
replicas on a data object with which a particular object is about to
interact. The function returns an error code if the targeted replica or
any of its siblings are intermediate or write-locked. The function was
adapted from an implementation of a local function inside of
rsDataObjOpen which made sure that replicas could not be opened without
a replica token if they were intermediate or at all if locked.