[#4236][#5421] Enforce ILL in unlink/rename/trim
These APIs now check the locked/intermediate status of a replica before
attempting to unlink/rename/trim it.

Also fixed an issue with rsDataObjRename where the FORCE_FLAG_KW was not
being honored, even though it seemed to be supported. This likely went
unnoticed because imv does not have a force flag option.