[#4331] Bumpers for data_object_finalize
Adds 2 features to the data_object_finalize API plugin:

 1. Checks permissions of the authenticated user to update system
    metadata on the specified data object.

 2. There are now a few immutable columns in r_data_main when attempting
    to modify through this interface. These include the data_id,
    data_name, and coll_id. These are logical level information which
    should not be changed outside of a specific operation (e.g. rename).

In order for these to work properly, logical locking and RST's publish
mechanism had to change in order to elevate privileges for
administrative operations. The publish namespace was added to RST and
the RST interface was changed to accept a publish::context struct. This
struct contains information like the RST entry key, replica identifier,
fileModified input, and a flag for elevating privileges. More can be
added in the future without impacting the interface too much.