[#5657] Fix a few memleaks
Mostly fixes memleaks concerning condInputs which should be free'd using
clearKeyVal. However, this also addresses a memleak found in the
dataObjInfo for the L1 descriptor when creating new replicas. The
dataObjInfo in the L1 descriptor now only refers to the opened replica
and any linked list information is free'd in rsDataObjOpen.