[#2719][#4331] Add tickets to data_object_finalize
Adds a new keyword to data_object_finalize input: bytes_written
This allows the caller to specify how many bytes were written in the
operation. This may have general applicability in the future but was
placed here to support updating the number of bytes written for tickets,
which always happens to be the entire size of the data object which was
written.

Adds a couple of new "high level ICAT" functions which allow for
updating ticket information outside of the database plugin. These call
into two new database operations - one to check permissions with support
for tickets (also updates ticket information except for bytes written)
and especially to update bytes written.

As a result of this change, new error handling capabilities have become
available from data_object_finalize. The output string contains a JSON
structure with an error message and an indication of whether the
database was updated successfully. In this way, users can determine the
appropriate action to take based on the nature of the failure.