[#5017] Set dataSize for parallel copy
The dataObjInp.dataSize of the opened L1 descriptor for the destination
data object is expected to have the data size of the source data object
in order for parallel transfer to correctly set the number of threads
for the portal in a federated transfer. This was lost somewhere along
the way and now has been restored so that copies across federation can
correctly transfer data using the portals put in place by the legacy
parallel transfer mechanism.

Copy does not need to resolve the resource hierarchy for the destination
replica as this can be done in rsDataObjOpen. This logic was here mainly
to uphold the OVERWRITE_WITHOUT_FORCE_FLAG rules for icp, but this is
being handled earlier in the function and the deleted logic was only
checking for existing replicas anyway.