[#5017] Prevent coll mtime updates for remote zone
When an open or put is being done over federation, the remote zone
collection mtime does not need to be updated because this logic should
be taken care of in the remote zone's API implementation. If there is no
collection mtime updating logic in that zone, the collection mtime being
updated is not important to that version of iRODS and so it shouldn't be
updated in this case either.

This avoids a known limitation in the iRODS protocol which disallows
further communication with the connected server in the remote zone from
the server in the local zone after a portal has been established for a
parallel transfer and the information about the portal is supposed to be
communicated back to the connected client in order to establish a direct
connection with the server in the remote zone.