[_XXXX] add DB plugin op for data_object_finalize
This adds a database plugin operation for use in the
data_object_finalize API plugin. This is meant to act as interchangeable
logic with the nanodbc-based database interactions which exists in the
API plugin today. This change is needed because of some possible bugs in
the implementation and/or in the nanodbc library itself. For more
details on the issue, see: https://github.com/nanodbc/nanodbc/issues/267