[_4094] Fix acPostProcForPut test
The test test_acPostProcForPut_replicate_to_multiple_resources was
operating on the assumption that replicating to an existing, marked-good
replica was allowed and does not produce an error. This has changed, and
the test has succeeded because the errors from acPostProcForPut were not
propagated to the iput which triggered it. In fact, the acPostProcForPut
which failed was being triggered by rsDataObjRepl being called from the
acPostProcForPut triggered by the iput. So, the operation attempts to
replicate over itself and SYS_NOT_ALLOWED is returned. This issue was
uncovered by the test running with the Python REP configured.

Now, the acPostProcForPut action for this test is only triggered by
actual PUT operations.

This change also fixes a syntax error in the Python implementation of
the rule text for this test.