[#5675] Set L1desc input data size for put
The L1desc input data size is used by the traditional parallel transfer
portal logic in order to determine the number of threads to use in the
parallel transfer. In a forced overwrite via put, the size is 0 because
the truncate flag records a size of 0. This change sets the dataSize in
the L1desc input to the size given to the put API by the user in the
input.