[#5682] Query for missing replica info in create
When an operation requires that the connection be redirected to the
catalog consumer which owns the storage resource, the replica
information from the database operation to register the new replica in
the catalog is lost because the rsPhyPathReg interface has no way of
returning the information.

This change allows the logic for creating a new replica to detect the
situation in which the rsPhyPathReg call had to redirect to the catalog
provider and queries the database in order to retrieve the information
which was generated by the registration database operation.