[#5675] Set L1desc input data size for repl/phymv
The L1desc input data size is used by the traditional parallel transfer
portal logic in order to determine the number of threads to use in the
parallel transfer. In replication (and phymv), the dataObjInp of the L1
descriptor needs to have the dataSize much like the other operations
which leverage this functionality.