[#5691] Relax usage of default resource for put
When determining the resource which will be receiving data in a put
operation, the default resource should be considered when overwriting an
exiting data object, but only when a replica exists on the default
resource. Otherwise, the voting mechanism should be allowed to select
the resource which will be overwritten. In the future, a more surgical
approach may be required, but for now we will follow the historical
behavior.