[#5689] Limit redirection in data_object_finalize
In cases where the connected server is not the catalog service
provider, the call to the remote data_object_finalize should not attempt
to perform the file_modified operation, should it be requested in the
input. After the catalog has been updated in the CSP, the RPC should
return and the control should continue on the connected server to
execute the file_modified operation.

This change now explicitly turns off trigger_file_modified in the input
to the redirected call to data_object_finalize in order to only update
the catalog on the CSP (again, if it is not the connected server).