[#5733] Only commit for tickets in check_permission_to_modify_data_object
The only time the database should commit as a result to a call to the
function cmlCheckDataObjId is for tickets. The DB operation
db_check_permission_to_modify_data_object_op used by the finalize API
calls commit every time, which results in a warning message in the
database log for postgres about committing while no transaction is in
progress. This change adds a check for the presence of ticket
information before committing or rolling back in this DB operation. This
is not necessarily a permanent solution as the warning message will
still appear if the ticket information was not actually updated as a
result of this call.