[#4977] Add Cron Facility and Agent Factory and Delay Server restarting
iRODS forks processes to handle the various services it provides, and because of this
it is possible to cause them to crash.

So now there is a facility to scheduling tasks to be executed at a given interval.
So now it checks to make sure the agent factory and the delay server exist, and should continue
to exist.