[#5759] Replace rsPhyPathReg in rsDataObjOpen
The new register_physical_path API plugin provides a similar interface to
perform the same task as rsPhyPathReg but also an output buffer to return
information about the registered replica back to the caller. The use of
rsPhyPathReg has been replaced in rsDataObjOpen in favor of this new API
plugin. This allows us to bypass the extra query performed after
registration to fill in the missing information about the registered
replica after registration has occurred due to the lack of information
passed back to the caller in the old interface.