[#5774] Propagate errors from touch API on CSC
When connected to a Catalog Service Consumer, the touch API was failing
to propagate errors resulting from failures back to the client. This was
caused by an unnecessary redirection in the touch API plugin to the
catalog service provider. The new established client connection
maintains a different rError stack, so the error messages were not
making it back to the client.