[#5848] Check target replica in write-lock check
When ill::try_lock checks the replica status for a replica when
determining whether or not it is allowed to be open, the replica itself
cannot be write-locked and if it is intermediate, the replica token must
match. This fixes a bug where opening an intermediate replica with a
valid replica token was not working when a second replica exists that is
write-locked (as it should be when there is an intermediate replica).