[#5863] Update mtime for creates/non-empty writes
Historically, the mtime for a replica was only updated if it was opened
for write (that is, it must have already existed) and bytes were written
during the operation. This changes the logic such that any operation
that results in a create or a bytes-writing open-for-write will update
the mtime for the target replica. This means that the mtime and create
time for a new replica may not be equal.