[#5851] Forward repl ID in make_dstream_factory
The ReplicaId passed to make_dstream_factory is being captured by
reference in the returned lambda, but this has high potential for a
dangling reference depending on how the function is used. This is
especially relevant for the leaf_resource_name and root_resource_name
types. Here is an example:

  make_dstream_factory(comm, path, io::leaf_resource_name{"lol"})

The leaf_resource_name will go out of scope and its value member will be
freed before the factory can be used to construct a stream. For now, we
should just use std::forward to forward the parameter to the lambda so
that it has ownership over its lifetime.