[#5894] Remove cross-zone copy case in close
The cross-zone copy operation in rsDataObjClose handles the legacy case
of copying a data object to a remote zone by registering it in the
catalog as part of finalizing the object. This is no longer needed in
4.2.9+ and in fact breaks federation for small file copies from a zone
running an iRODS server version before 4.2.9 to a remote zone running an
iRODS server version 4.2.9 or later.

The client API whitelist should also expose the L3 file single buffer
put API as it is used in the legacy rsDataObjCopy (4.2.8 and earlier)
and so is needed in order for federation to work.