[#5917][#5923] Validate server negotiation keys
The negotiation_key is a way of encrypting the zone_key in negotiations
for server-to-server connections. The negotiation_key is used to sign
the zone_key sent to the remote server to perform the "advanced
negotiation" step and determine how the rest of the communications will
be performed. The documentation states that the negotiation_key must be
defined and exactly 32 alphanumeric (plus underscores) characters.
Furthermore, the negotiation_key must be the same across a zone. This
change ensures that these requirements are being enforced.