[#5950] Add rsComm_t members to RE serializer
The RE serialization for an rsComm_t* does not include certain members
which were available by default as session variables in static PEPs. In
order to support dynamic PEP use cases, these are now serialized for use
in dynamic PEPs which have an rsComm_t* in the signature. The following
members are now being serialized:
  - connectCnt
  - sock
  - option
  - status
  - apiInx