[#5950] Add test for rsComm_t* RE serialization
Adds a test which implements a PEP with an rsComm_t* in the signature
and checks to see whether various members of the struct are present and
are equal to the expected value.

Adds a unit test which tests the serialization of an RsComm* when it is
a nullptr, a 0-initialized structure, and a fully populated structure.