[#5700] Control plane resolves hostname aliases using hosts_config.json.
The important change in this commit is the introduction and use of the
function, resolve_hostnames_using_hosts_config.

This new function is called after parsing the request. The function
takes the incoming list of hostnames and replaces each one with the
resolved hostname from hosts_config.json. Hostnames that do not have an
entry in hosts_config.json remain in the hostname list as-is.

This commit also cleans up formatting and simplifies logic.