[#6019] Implemented new allocator: capped_memory_resource
capped_memory_resource is an alternative PMR allocator to
fixed_buffer_resource. It exists due to a bug in Boost.Container's PMR
implementation.

While capped_memory_resource does not match the level of performance of
fixed_buffer_resource, it does satisfy the requirement of enforcing a
limit on the number of bytes that can be allocated through it.