[#4649] Fix leaks in fill msparam functions
The functions for filling the inOutStruct for msParams for various types
can leak previously allocated inOutStructs as well as the type
parameters. The msParam structures are often reused for both in and out
parameters when calling rules. This will require ensuring that
previously free'd inOutStruct pointers are set to nullptr.