[#5927] Add support for checksum keywords in repl
This change adds support for 3 mutually exclusive keywords in rsDataObjRepl:

  1. VERIFY_CHKSUM_KW - If provided, a checksum is computed for the
destination replica and compared with the checksum on the source replica
to verify that the data arrived correctly. If the source replica does
not have a checksum, an error occurs.

  2. REG_CHKSUM_KW - If provided, a checksum is computed for the
destination replica and is not compared with the checksum on the source
replica (analogous to "forceChksum" in microservices). If the source
replica has a checksum, an error occurs because an inconsistent state
can be created by having two marked-good replicas with differing
checksums.

  3. NO_COMPUTE_KW - If provided, a checksum will not be computed for
the destination replica even if the source replica has a checksum. This
can also create an inconsistent state by having two marked-good replicas
with differing contents, but the user has requested that the
verification level of the replication be downgraded to a size check.