[#6089] Prevent checksums on archive replicas
Archive resource plugins often do not have the ability to checksum data
because it is expensive or is simply unsupported. Before this change,
iRODS would take the checksum from the replica in cache and trusts that
it is correct. The `NO_COMPUTE_KW` can now be passed to `rsDataObjClose`
and the finalize logic for replication operations will not calculate a
checksum.