[#6114] Use list comprehension for suite filter
filter() is a generator in python3 which makes the "marker" filter
used for running specific tests much harder to use the same way in
python3. In python2, a filter was executed immediately and converted
into a list, so a list comprehension is used instead here.
Unfortunately, the variable cannot be seen inside of the nested function
which uses it, so it is now passed as an argument.