[#5479] Add authentication API plugin and framework
This change adds a new authentication plugin framework. The basis for
this work was originally developed for the Authentication Working Group
and its implementation and documentation can be found here:

  https://github.com/irods-contrib/irods_working_group_authentication

The framework provides a single function called authenticate_client
which allows a connected client user to be authenticated. The only
plugin which has been implemented using the new framework so far is the
iRODS native authentication. The legacy authentication plugin will
remain in place in order to support talking to older servers with newer
clients and servers.

This change also adds a new API plugin which loads an auth plugin and
calls the operation specified by "next_operation" in the JSON input.
This is used by the authentication plugins within the framework to make
requests and receive responses from the server.

The new authentication plugin framework has been integrated with
clientLogin so that all authentication requests are routed through the
new framework. Authentication requests for clients connected to servers
before 4.3.0 will use the older authentication flow.

As a result of using the authentication framework, a new requirement for
server startup has been introduced: API plugins must be loaded at server
startup in order for the server to start. So, part of this change also
loads the client and server API tables on server startup so that any
server startup operations which require API plugins can be performed.
The delay server also requires that the client API table be loaded on
startup so that any API plugins required for communication with the
server can be loaded.